Reorganize backend, create routes folder and auth.js file for authentication routes, modify index file to only manage express server
