+++ backend/routes/auth.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+const authController = require("../controllers/authController.js");
+
+router.post("/register", authController.register);
+router.post("/login", authController.login);
+
+module.exports = router;
