+++ frontend/src/Login.jsx
@@ -12,7 +12,7 @@ function Login() {
 
     const handleSubmit = (e) => {
         e.preventDefault();
-        axios.post('http://localhost:3001/login', { email, password })
+        axios.post('http://localhost:3001/auth/login', { email, password })
             .then(result => {
                 console.log(result);
                 console.log('Response data:', result.data);
